+++ src/views/LogsPage.vue
@@ -15,7 +15,7 @@
         <div class="card mb-1 block break-all p-2 text-sm">
           <span>{{ item.seq }}</span>
           <span class="mx-2 text-primary">
-            {{ dayjs(item.time).locale(language).format('HH:mm:ss') }}
+            {{ item.time }}
           </span>
           <span :class="textColorMapForType[item.type as keyof typeof textColorMapForType]">
             {{ item.type }}
@@ -32,9 +32,7 @@ import VirtualScroller from '@/components/common/VirtualScroller.vue'
 import { LOG_LEVEL } from '@/config'
 import { isMiddleScreen } from '@/helper/utils'
 import { logFilter, logs } from '@/store/logs'
-import { language } from '@/store/settings'
 import type { LogWithSeq } from '@/types'
-import dayjs from 'dayjs'
 import { computed } from 'vue'
 
 const textColorMapForType = {
@@ -45,12 +43,14 @@ const textColorMapForType = {
 }
 
 const renderLogs = computed(() => {
-  return logs.value.filter((log) => {
-    if (logFilter.value) {
-      return log.payload.includes(logFilter.value) || log.type.includes(logFilter.value)
-    }
+  if (logFilter.value) {
+    const regex = new RegExp(logFilter.value, 'i')
 
-    return true
-  })
+    return logs.value.filter((log) => {
+      return [log.payload, log.time, log.type].some((i) => regex.test(i))
+    })
+  }
+
+  return logs.value
 })
 </script>